leetcode/src/main/java/org/issxjl2015/algorithm/Lc_448_找到所有数组中消失的数字/Solution.java
package org.issxjl2015.algorithm.Lc_448_找到所有数组中消失的数字;

import java.util.ArrayList;
import java.util.HashSet;
import java.util.List;
import java.util.Set;

public class Solution {

    public List<Integer> findDisappearedNumbers(int[] nums) {
        Set<Integer> set = new HashSet<>();
        List<Integer> list = new ArrayList<>();

        for (int num : nums) {
            set.add(num);
        }

        for (int i = 1; i <= nums.length; i++) {
            if (!set.contains(i)) {
                list.add(i);
            }
        }
        return list;
    }

}
